Add files via upload
The scripts zip folder contains the code for my Assignment. I have also included a zipped folder for my assets (couldn't do the whole project) as I did not know what was desired, there is a thirdparty download in there but this was only for material skins which were not used in the end.